# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentação da GateWay API GymApp",
    "description": "Documentação da gateway api de suporte à aplicação GymApp",
    "contact": { "name": "Teste", "email": "teste@teste.com" },
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:2900", "description": "API de teste" }
  ],
  "paths": {
...
    "/admin/marca/": {
      "post": {
        "summary": "Registo de marcas",
        "description": "Esta rota vai registar na base de dados uma marca associada a um admin",
        "tags": ["Marca"],
        "security": [{ "basicAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "nome": { "type": "string" },
                  "mobilidade": { "type": "boolean" },
                  "cor": { "type": "string" },
                  "logotipo": { "type": "string" }
                }
              },
              "examples": {
                "Inserir Marca": {
                  "value": {
                    "nome": "Galopa",
                    "mobilidade": true,
                    "cor": "vermelho",
                    "logotipo": "teste"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Inserir Marca": {
                    "value": { "msg": "Marca inserida com sucesso" }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Ver todas as marcas de um admin",
        "description": "Esta rota vai apresentar todas as marcas de um admin",
        "tags": ["Marca"],
        "security": [{ "basicAuth": [] }],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Ver Marcas de um Admin": {
                    "value": { "msg": "Operação realizada com sucesso" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/ginasio/editar": {
      "put": {
        "summary": "Editar ginasio",
        "description": "Esta rota vai editar um ginasio criado",
        "tags": ["Marca"],
        "security": [{ "basicAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "ginasioId": { "type": "string" },
                  "nome": { "type": "string" },
                  "rua": { "type": "string" },
                  "descricao": { "type": "string" },
                  "imagemUrl": { "type": "boolean" },
                  "lat": { "type": "boolean" },
                  "long": { "type": "boolean" }
                }
              },
              "examples": {
                "Editar ginasio": {
                  "value": {
                    "ginasioId": "059f2ecf-cdcc-408c-a139-1ceb7df0742e",
                    "nome": "Ginásio 4",
                    "rua": "Rua das Castanhas",
                    "descricao": "null",
                    "imagemUrl": "teste",
                    "lat": "0",
                    "long": "0"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Editar ginasio": {
                    "value": {
                      "ginasio_id": "531316a4-b454-42ec-bc01-7358a8c3bc48",
                      "nome": "Ginásio Teste",
                      "rua": "Rua das Castanhas",
                      "cp_id": "f1d6a939-abde-40c9-aa17-97916e3d2b94",
                      "marca_id": "f34fdd41-8670-4b50-9311-ebdb0ad993c4",
                      "tag": "G4_01",
                      "descricao": "null",
                      "imagem_url": "teste",
                      "lat": "0",
                      "long": "0",
                      "isDeleted": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/marca/ginasio/": {
      "post": {
        "summary": "Registo de ginasios",
        "description": "Esta rota vai registar na base de dados um ginasio que esta associado a uma marca",
        "tags": ["Marca"],
        "security": [{ "basicAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "marcaId": { "type": "string" },
                  "nome": { "type": "string" },
                  "rua": { "type": "string" },
                  "cp": { "type": "number" },
                  "cpExt": { "type": "number" },
                  "imagemUrl": { "type": "string" },
                  "lat": { "type": "string" },
                  "long": { "type": "string" }
                }
              },
              "examples": {
                "Inserir Registo de Ginásios": {
                  "value": {
                    "marcaId": "2c4fc500-7373-44ed-a665-270f19da455c",
                    "nome": "Ginásio 4",
                    "rua": "Rua das Castanhas",
                    "cp": 3850,
                    "imagemUrl": "3",
                    "lat": "0",
                    "long": "0",
                    "cpExt": 821
                  }
                }
              }
            }
          }
        },
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Inserir Agendamento": {
                    "value": {
                      "ginasio_id": "cca94698-1048-487e-a67f-e238f2edb9d3",
                      "nome": "Ginásio 4",
                      "rua": "Rua das Castanhas",
                      "cp_id": "867277a8-8f5a-4148-a028-32621a4a1a48",
                      "marca_id": "d9017634-c824-423d-96a3-6da7f162917a",
                      "tag": "G4_03",
                      "descricao": null,
                      "imagem_url": "teste",
                      "lat": "0",
                      "long": "0",
                      "isDeleted": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/marca/{marcaId}/ginasio/": {
      "get": {
        "summary": "Ver todos os ginásios de uma marca",
        "description": "Esta rota vai apresentar todos os ginásios de uma marca",
        "tags": ["Marca"],
        "security": [{ "basicAuth": [] }],
        "parameters": [
          {
            "name": "marcaId",
            "in": "path",
            "description": "Id da marca",
            "required": true,
            "schema": { "properties": { "marcaId": { "type": "string" } } }
          }
        ],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Ver Ginásios de uma Marca": {
                    "value": { "msg": "Operação realizada com sucesso" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/ginasio/modalidades": {
      "post": {
        "summary": "Registo das modalidades do ginásio",
        "description": "Esta rota vai registar na base de dados uma modalidade de um ginásio em específico",
        "tags": ["Modalidades"],
        "security": [{ "basicAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "ginasioId": { "type": "string" },
                  "nome": { "type": "string" },
                  "imagemUrl": { "type": "string" }
                }
              },
              "examples": {
                "Inserir Modalidade": {
                  "value": { "ginasioId": "059f2ecf-cdcc-408c-a139-1ceb7df0742e","nome": "Caminhada", "imagemUrl": "teste" }
                }
              }
            }
          }
        },
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Inserir Modalidade": {
                    "value": {
                      "modalidade_id": "a7204142-de0b-4c5e-915f-7905109612c5",
                      "ginasio_id": "531316a4-b454-42ec-bc01-7358a8c3bc48",
                      "nome": "Musculação",
                      "imagem_url": "http://musculacao.imagem",
                      "isDeleted": false
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Editar modalidades",
        "description": "Esta rota vai editar uma modalidade de um ginásio em específico",
        "tags": ["Modalidades"],
        "security": [{ "basicAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "ginasioId": { "type": "string" },
                  "modalidadeId": { "type": "string" },
                  "nome": { "type": "string" },
                  "imagemUrl": { "type": "string" }
                }
              },
              "examples": {
                "Editar Exercicio": {
                  "value": {
                    "ginasioId": "059f2ecf-cdcc-408c-a139-1ceb7df0742e",
                    "modalidadeId": "03edb52e-6b99-49d3-ab26-0504a7ebf6b4",
                    "nome": "Crossfit",
                    "imagemUrl": "https://crossfit.imagem",
                    "isDeleted": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Editar Modalidade": {
                    "value": { "msg": "Modalidada editada com sucesso" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/ginasio/{ginasioId}/modalidades": {
      "get": {
        "summary": "Ver todas as modalidades de um ginásio",
        "description": "Esta rota vai apresentar todas as modalidades de um ginásio",
        "tags": ["Modalidades"],
        "security": [{ "basicAuth": [] }],
        "parameters": [
          {
            "name": "ginasioId",
            "in": "path",
            "description": "Id da ginasio",
            "required": true,
            "schema": { "properties": { "ginasioId": { "type": "string" } } }
          }
        ],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Ver Modalidades de um Ginásio": {
                    "value": [
                      {
                        "nome": "Musculação",
                        "imagem_url": "http://musculacao.imagem"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/aluno/remover/{uId}": {
      "delete": {
        "summary": "Remover aluno",
        "description": "Esta rota vai remover um aluno",
        "tags": ["Marca"],
        "security": [{ "basicAuth": [] }],
        "parameters": [
          {
            "name": "uId",
            "in": "path",
            "description": "Id do aluno",
            "required": true,
            "schema": { "properties": { "uId": { "type": "string" } } }
          }
        ],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } }
              }
            }
          }
        }
      }
    },
    "/admin/ginasio/{ginasioId}/modalidades/{modalidadeId}": {
      "delete": {
        "summary": "Remover modalidades",
        "description": "Esta rota vai apagar na base de dados uma modalidade de um ginásio em específico",
        "tags": ["Modalidades"],
        "security": [{ "basicAuth": [] }],
        "parameters": [
          {
            "name": "modalidadeId",
            "in": "path",
            "description": "Id da modalidade",
            "required": true,
            "schema": { "properties": { "modalidadeId": { "type": "string" } } }
          },
          {
            "name": "ginasioId",
            "in": "path",
            "description": "Id do ginasio",
            "required": true,
            "schema": { "properties": { "modalidadeId": { "type": "string" } } }
          }
        ],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Remover Modalidade": {
                    "value": {
                      "modalidade_id": "047ea26d-bc43-4463-a3d1-8ea726a27b9f",
                      "ginasio_id": "dc8acc46-89eb-4d0f-a14a-2388b21c90a0",
                      "nome": "Corrida",
                      "imagem_url": "https://corrida.imagem",
                      "isDeleted": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/treinador/{treinador_id}": {
      "delete": {
        "summary": "Remover treinador",
        "description": "Esta rota vai remover um treinador",
        "tags": ["Marca"],
        "security": [{ "basicAuth": [] }],
        "parameters": [
          {
            "name": "treinador_id",
            "in": "path",
            "description": "Id do treinador",
            "required": true,
            "schema": { "properties": { "treinador_id": { "type": "string" } } }
          }
        ],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Remover Treinador": {
                    "value": { "msg": "Treinador cessado das funções" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/marca/{marcaId}": {
      "delete": {
        "summary": "Remover marca",
        "description": "Esta rota vai remover uma marca de um ginásio",
        "tags": ["Marca"],
        "security": [{ "basicAuth": [] }],
        "parameters": [
          {
            "name": "marcaId",
            "in": "path",
            "description": "Id da marca",
            "required": true,
            "schema": { "properties": { "marcaId": { "type": "string" } } }
          }
        ],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Remover Marca": {
                    "value": { "msg": "Marca removida com sucesso" }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Ver a informação de uma marca",
        "description": "Esta rota vai apresentar a informação de uma marca",
        "tags": ["Marca"],
        "security": [{ "basicAuth": [] }],
        "parameters": [
          {
            "name": "marcaId",
            "in": "path",
            "description": "Id da marca",
            "required": true,
            "schema": { "properties": { "marcaId": { "type": "string" } } }
          }
        ],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Ver Ginasio": {
                    "value": { "msg": "Operação realizada com sucesso" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/marca/espf": {
      "put": {
        "summary": "Editar Marca",
        "description": "Esta rota vai editar uma Marca criada",
        "tags": ["Marca"],
        "security": [{ "basicAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "marcaId": { "type": "string" },
                  "nome": { "type": "string" },
                  "cor": { "type": "string" },
                  "logotipo": { "type": "string" },
                  "mobilidade": { "type": "boolean" },
                  "isDeleted": { "type": "boolean" }
                }
              },
              "examples": {
                "Editar Marca": {
                  "value": {
                    "marcaId": "2c4fc500-7373-44ed-a665-270f19da455c",
                    "nome": "Galo",
                    "cor": "vermelho",
                    "logotipo": "teste1",
                    "mobilidade": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Editar Marca": {
                    "value": { "msg": "Marca editada com sucesso" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/marca/treinadores": {
      "post": {
        "summary": "Registar Treinadores",
        "description": "Esta rota permite registar treinadores numa marca",
        "tags": ["Marca"],
        "security": [{ "basicAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "marcaId": { "type": "string" },
                  "email": { "type": "string" },
                  "nome": { "type": "string" },
                  "password": { "type": "string" },
                  "data_nasc": { "type": "string" },
                  "data_entrada": { "type": "string" },
                  "genero": { "type": "number" }
                }
              },
              "examples": {
                "Registar Treinador": {
                  "value": {
                    "marcaId": "2c4fc500-7373-44ed-a665-270f19da455c",
                    "email": "treinador@treinador.pt",
                    "nome": "treinador",
                    "password": "treinador",
                    "dataNasc": "2002-04-09",
                    "dataEntrada": "2022-03-23",
                    "genero": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Treinador"
                },
                "examples": {
                  "Registar Treinador": {
                    "value": {
                      "treinador": {
                        "uid": "919dc8c8-9853-4b13-b0e7-9b88fe587137",
                        "email": "treinador1@admin.com",
                        "nome": "treinador",
                        "password": "$2b$08$y09XOo8D9OmbYUhoNkuOFejHGPAeRtEJu5yhwnofaDHw4lmgPvyby",
                        "data_nasc": "2002-04-09T00:00:00.000Z",
                        "hashtag": "t_01",
                        "data_entrada": "2022-03-23T00:00:00.000Z",
                        "funcao_id": "4c1c8eaf-7207-4d06-864a-a50191e9f5a7",
                        "refresh_token": null,
                        "genero": 0,
                        "pontos": 0,
                        "descricao": null,
                        "imagem_url": null,
                        "isDeleted": false
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/notificacao/user": {
      "post": {
        "summary": "Notificação para user",
        "description": "Esta rota permite enviar uma notificação para um user",
        "tags": ["Notificações"],
        "security": [{ "basicAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "destinoId": { "type": "string" },
                  "conteudo": { "type": "string" },
                  "tipo": { "type": "number" }
                }
              },
              "examples": {
                "Criar Notificação": {
                  "value": { "destinoId": "000d1e14-617e-423e-8a1a-f63d4fa5af6a", "conteudo": "Teste", "tipo": 0 }
                }
              }
            }
          }
        },
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/User"
                },
                "examples": {
                  "Criar Notificação": {
                    "value": {
                      "noti_id": "4363ff26-e2ae-426b-aaa9-6c9524858e56",
                      "origem_uid": "332b256d-d646-4877-8a4f-f5538472e0d4",
                      "conteudo": "Teste5",
                      "data": "2022-04-18T17:40:21.382Z",
                      "tipo": 0
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": { "basicAuth": { "type": "http", "scheme": "bearer" } },
    "schemas": {
...
      "Treinador": {
        "type": "object",
        "properties": {
          "uid": { "type": "string" },
          "email": { "type": "string" },
          "nome": { "type": "string" },
          "password": { "type": "string" },
          "data_nasc": { "type": "string" },
          "hashtag": { "type": "string" },
          "data_entrada": { "type": "string" },
          "funcao_id": { "type": "string" },
          "refresh_token": { "type": "string" },
          "genero": { "type": "number" },
          "pontos": { "type": "number" },
          "descricao": { "type": "string" },
          "imagem_url": { "type": "string" },
          "isDeleted": { "type": "string" }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "uid": { "type": "string" },
          "email": { "type": "string" },
          "nome": { "type": "string" },
          "password": { "type": "string" },
          "data_nasc": { "type": "string" },
          "hashtag": { "type": "string" },
          "data_entrada": { "type": "string" },
          "funcao_id": { "type": "string" },
          "refresh_token": { "type": "string" },
          "genero": { "type": "number" },
          "pontos": { "type": "number" },
          "descricao": { "type": "string" },
          "imagem_url": { "type": "string" },
          "isDeleted": { "type": "boolean" }
        }
      },
...
    }
  }
}